feat: Move React hooks to react-utils package
