+++ packages/react-utils/src/debug-context/Context.tsx
@@ -1,5 +1,5 @@
+import { assert, isNonEmptyString } from '@contember/utilities'
 import { ReactNode, createContext, memo, useContext, useMemo, useRef } from 'react'
-import { assert, isNonEmptyString } from '../assert-types'
 import { isNoopScopedConsole, noopScopedConsole } from './Constants'
 import { ScopedConsoleContextType } from './Types'
 
@@ -62,7 +62,7 @@ export const useScopedConsoleRef = (prefix: string) => {
 	return ref
 }
 
-type DebugChildrenProps =
+export type DebugChildrenProps =
 	| { active: true; children: ReactNode; id: string; }
 	| { active?: false; children: ReactNode; id?: string; }
