+++ packages/react-utils/src/hooks/unwrapRefValue.ts
@@ -4,7 +4,7 @@ export type RefObjectOrElement<T> = T | RefObject<T>
 
 /**
  * Unwraps a ref object or returns the value if it is not a ref object.
- * @param value Value to unwrap
+ * @param value - Value to unwrap
  * @returns Unwrapped value
  */
 export function unwrapRefValue<T>(value: RefObjectOrElement<T>): T | null {
