+++ packages/react-utils/src/hooks/useComposeRef.ts
@@ -1,7 +1,7 @@
 import { MutableRefObject, RefCallback, useMemo, useRef } from 'react'
 
-type MaybeRef<T> = ((instance: T | null) => void) | MutableRefObject<T | null> | null | undefined
-type ComposedRefCallback<T> = RefCallback<T> & { current: T | null };
+export type MaybeRef<T> = ((instance: T | null) => void) | MutableRefObject<T | null> | null | undefined
+export type ComposedRefCallback<T> = RefCallback<T> & { current: T | null };
 
 function setRef<T>(ref: MaybeRef<T>, instance: T | null): void {
 	if (ref) {
@@ -19,7 +19,7 @@ function setRef<T>(ref: MaybeRef<T>, instance: T | null): void {
  * Useful when you need to use outer ref and inner in the same time, e.g. passing element
  * back to forwardedRef an also using the same value within the component locally.
  *
- * @param refs Rest parameter of refs to fill with instance
+ * @param refs - Rest parameter of refs to fill with instance
  * @returns Returns a function similar to a RefCallback
  */
 export function useComposeRef<T>(...refs: MaybeRef<T>[]): ComposedRefCallback<T> {
