+++ packages/react-utils/src/hooks/useScrollOffsets.ts
@@ -1,7 +1,7 @@
-import { useReferentiallyStableCallback } from '@contember/react-utils'
+import { assert } from '@contember/utilities'
 import { useLayoutEffect, useMemo, useState } from 'react'
-import { assert } from '../assert-types'
-import { useScopedConsoleRef } from '../ui-debug-context'
+import { useScopedConsoleRef } from '../debug-context'
+import { useReferentiallyStableCallback } from '../referentiallyStable'
 import { RefObjectOrElement, unwrapRefValue } from './unwrapRefValue'
 import { useOnScrollWithin } from './useOnScrollWithin'
 import { useOnWindowResize } from './useOnWindowResize'
