+++ packages/react-utils/src/index.ts
@@ -1,3 +1,5 @@
+export * from './context'
+export * from './debug-context'
 export * from './hooks'
 export * from './referentiallyStable'
 export * from './types'
