+++ packages/react-utils/src/tsconfig.json
@@ -2,5 +2,8 @@
 	"extends": "../../../tsconfig.settings.json",
 	"compilerOptions": {
 		"outDir": "../dist/types"
-	}
+	},
+	"references": [
+		{ "path": "../../utilities/src" },
+	]
 }
