+++ packages/react-utils/src/types/PropsWithRequiredChildren.ts
@@ -1,0 +1,3 @@
+import { ReactNode } from 'react'
+
+export type PropsWithRequiredChildren<P = unknown> = P & { children: ReactNode | undefined };
