+++ packages/react-utils/src/types/Serializable.ts
